reorganized dir structure. Fixed things according to comments on PR
