+++ woodwork/statistics_utils/_infer_temporal_frequencies.py
@@ -1,9 +1,8 @@
-import pandas as pd
-
 from woodwork.logical_types import Datetime, Timedelta
 
+from woodwork.statistics_utils.frequency_inference._infer_frequency import infer_frequency
 
-def _infer_temporal_frequencies(dataframe, temporal_columns=None):
+def _infer_temporal_frequencies(dataframe, temporal_columns=None, debug=False):
     """Infers the observation frequency (daily, biweekly, yearly, etc) of each temporal column
             in the DataFrame. Temporal columns are ones with the logical type Datetime or Timedelta.
 
@@ -12,7 +11,8 @@ def _infer_temporal_frequencies(dataframe, temporal_columns=None):
         temporal_columns (list[str], optional): Columns for which frequencies should be inferred. Must be columns
             that are present in the DataFrame and are temporal in nature. Defaults to None. If not
             specified, all temporal columns will have their frequencies inferred.
-
+        debug (boolean): A flag used to determine if more information should be returned for each temporal column if
+                no uniform frequency was found.
     Returns:
         (dict): A dictionary where each key is a temporal column from the DataFrame, and the
             value is its observation frequency represented as a pandas offset alias string (D, M, Y, etc.)
@@ -36,4 +36,4 @@ def _infer_temporal_frequencies(dataframe, temporal_columns=None):
                     f"Cannot determine frequency for column {col} with logical type {ltype}"
                 )
 
-    return {col: pd.infer_freq(dataframe[col]) for col in temporal_columns}
+    return {col: infer_frequency(observed_ts=dataframe[col], debug=debug) for col in temporal_columns}
