+++ woodwork/statistics_utils/frequency_inference/_determine_most_likely_freq.py
@@ -1,7 +1,7 @@
 from ._constants import FREQ_INFERENCE_THRESHOLD, NON_INFERABLE_FREQ
 
 
-def _determine_most_likely_freq(alias_dict):
+def _determine_most_likely_freq(alias_dict, threshold=FREQ_INFERENCE_THRESHOLD):
 
     alias_info = alias_dict.values()
 
@@ -11,7 +11,7 @@ def _determine_most_likely_freq(alias_dict):
     most_likely_alias = sorted_aliases[0]["alias"]
     most_likely_count = sorted_aliases[0]["count"]
 
-    if most_likely_count / n_total < FREQ_INFERENCE_THRESHOLD:
+    if most_likely_count / n_total < threshold:
         return None
 
     if most_likely_alias is not NON_INFERABLE_FREQ:
