+++ woodwork/statistics_utils/frequency_inference/_generate_freq_candidates.py
@@ -1,12 +1,13 @@
 import pandas as pd
-from ._constants import NON_INFERABLE_FREQ, WINDOW_LENGTH, CANDIDATE_COLUMN_NAME
+from ._constants import NON_INFERABLE_FREQ, WINDOW_LENGTH
 
 
-def _generate_freq_candidates(time_series):
+def _generate_freq_candidates(time_series, window_length=WINDOW_LENGTH):
     """Calculate a set of candidate frequecies for incoming time_series
 
     Args:
         time_series (Series): The time_series for which candidate frequencies will be calculated.
+        window_length (int): The length of the window. Default is 15
 
     Returns:
         (dict[pandas_alias(str) -> dict]): Returns a dictionary where each key is candidate Pandas Frequency Alias and the value is
@@ -18,8 +19,8 @@ def _generate_freq_candidates(time_series):
     """
     
     alias_dict = {}
-    for window in time_series.rolling(WINDOW_LENGTH):
-        if len(window) == WINDOW_LENGTH:
+    for window in time_series.rolling(window_length):
+        if len(window) == window_length:
 
             # calculate alias 
             alias = pd.infer_freq(window) or NON_INFERABLE_FREQ
@@ -30,7 +31,7 @@ def _generate_freq_candidates(time_series):
             if alias in alias_dict:
                 curr_alias = alias_dict[alias]
                 curr_alias["count"] += 1
-                curr_alias["max_dt"] = window.iloc[WINDOW_LENGTH-1]
+                curr_alias["max_dt"] = window.iloc[window_length-1]
 
             else:
                 alias_dict[alias] = {
